feat: separate the scheduler into its own standalone process that can be run independently from the API server. This approach provides several benefits: Eliminates resource leakage between processes Allows the API to run with multiple workers without scheduler duplication Simplifies monitoring and debugging of each component Improves overall system stability Enables independent scaling of API and scheduler components feat: provide Fastnear api key to be set as bearer fear: env.reply to duplicate message in stdout fix: local composer default timezone set to UTC
