+++ hub/api/v1/thread_routes.py
@@ -580,8 +580,7 @@ class RunCreateParamsBase(BaseModel):
     instructions: Optional[str] = Field(
         None,
         description=(
-            "Overrides the instructions of the assistant. "
-            "This is useful for modifying the behavior on a per-run basis."
+            "Overrides the instructions of the assistant. This is useful for modifying the behavior on a per-run basis."
         ),
     )
     tools: Optional[List[dict]] = Field(None, description="Override the tools the assistant can use for this run.")
@@ -829,8 +828,7 @@ def _run_agent(
                 if parent_run.parent_run_id:
                     raise HTTPException(
                         status_code=400,
-                        detail="Parent run cannot have a parent run. "
-                        "Parent run is already a child run of another run.",
+                        detail="Parent run cannot have a parent run. Parent run is already a child run of another run.",
                     )
 
                 parent_run.child_run_ids.append(run_id)
